requirements.txt
absl-py==0.2.2
anaconda-client==1.6.8
asn1crypto==0.24.0
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
clyent==1.2.2
cryptography==2.1.4
decorator==4.2.1
dlib==19.13.1
face-recognition==1.2.2
face-recognition-models==0.3.0
gast==0.2.0
grpcio==1.12.1
html5lib==0.9999999
idna==2.6
imutils==0.4.5
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
Markdown==2.6.11
nbformat==4.4.0
numpy==1.14.0
opencv-contrib-python==3.4.0.12
paho-mqtt==1.3.1
pandas==0.23.0
Pillow==5.1.0
protobuf==3.5.2.post1
pycparser==2.18
pymongo==3.6.1
pyOpenSSL==17.5.0
PyQt5==5.10.1
pyqt5-tools==5.9.0.1.2
pyserial==3.4
PySocks==1.6.7
python-dateutil==2.6.1
pytz==2017.3
pyzbar==0.1.7
requests==2.18.4
ruamel-yaml==0.15.35
scipy==1.1.0
sip==4.19.8
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.22
Werkzeug==0.14.1
win-inet-pton==1.0.1
wincertstore==0.2
